[Modern Find/Replace] Refactor FindReplaceDialog
The FindReplaceDialog's business-logic is now handled in the
FindReplacer.
The FindReplacer communicates using a FindReplaceStatus object which is
polled by FindReplaceDialog as needed and allows for displaying correct
messages in the Interface.

This Refactoring is required so that the business-logic can be reused in
the new FindReplace overlay, as proposed in issue eclipse-platform#1090