Fix display used for ImageRegistry creation
The AbstractUIPlugin, which is implemented by several UI plug-ins,
creates and provides an ImageRegistry. This registry is initialized with
the display of the current thread, if the current thread has a display,
and with the display of the workbench otherwise. In consequence,
if the first access to the ImageRegistry happens from a thread with a
(temporary) display that is disposed at some point in time, all images
will (unexpectedly) be disposed as well. This can currently only happen
on Windows, as Linux and MacOS do not allow multiple display.

This change adapts the ImageRegistry creation so that the display of the
workbench is used, if available. The current thread's display is only
used if no workbench is running. This results in the ImageRegistry being
disposed with the workbench instead of any temporary display. On systems
that have no support for multiple displays, the change will simply have
no effect. The commit also adds a regression test for the unexpected
behavior.