Move identification of find/replace operation success to status object
The recently introduced IFindReplaceStatus objects used by the
FindReplaceLogic to notify consumers (like the FindReplaceDialog) about
the result of a find/replace operation only provide the information
about the success of an operation implicitly. This is why the
FindReplaceDialog currently needs to consider the type of an
IFindReplaceStatus to extract this information.

With this change, the IFindReplaceStatus objects provide success
information via a `wasSuccessful()` method, which relieves the
FindReplaceDialog from manually distinguishing and processing different
types of status objects.