Event processing document re-write
Event_Processing used to be a working document for the design of the
Event Admin usasge in the Eclipse platform. As this document is linked
by lots of other document it has been updated to reflect the decisions
taken in the past, rather than posting these questions as something
still to be solved.

Also delete dead links and fix links to images